requirements.txt
kivy
kivymd
pymupdf
ebooklib
pyttsx3
plyer
Pillow
kivy_garden.cefpython